+++ airbng-frontend/src/pages/admin/AdminDashboardPage.jsx
@@ -1,0 +1,11 @@
+import AdminLayout from "../../components/admin/AdminLayout";
+
+
+function AdminDashboardPage() {
+    return (
+        <div>
+            <AdminLayout />
+        </div>
+    )
+}
+export default AdminDashboardPage;
