--- airbng-frontend/src/pages/admin/AdminHomePage.jsx
@@ -1,9 +1,0 @@
-function AdminHomePage() {
-    return (
-        <div>
-            <h1>Admin Home Page</h1>
-            <p>Welcome to the admin dashboard!</p>
-        </div>
-    );
-}
-export default AdminHomePage;
