+++ airbng-frontend/src/styles/admin/layout/AdminLayout.module.css
@@ -1,0 +1,23 @@
+.pageContainer {
+    width: 100%;
+    height: 100vh;
+    display: flex;
+    flex-direction: column;
+}
+
+.container {
+    display: flex;
+    height: calc(100vh - 60px);
+    background-color: #f8f9fa;
+}
+
+.mainContent {
+    flex: 1;
+    overflow-y: auto;
+}
+
+/*.contentWrapper {*/
+/*    padding: 30px;*/
+/*    max-width: 1200px;*/
+/*    margin: 0 auto;*/
+/*}*/
