+++ airbng-frontend/src/styles/admin/layout/AdminSidebar.module.css
@@ -1,0 +1,139 @@
+.sidebar {
+    background-color: #ffffff;
+    color: #333333;
+    transition: width 0.3s ease;
+    display: flex;
+    flex-direction: column;
+    border-right: 1px solid #e0e0e0;
+    box-shadow: 2px 0 4px rgba(0, 0, 0, 0.1);
+}
+
+.sidebarOpen {
+    width: 280px;
+}
+
+.sidebarClosed {
+    width: 80px;
+}
+
+.sidebarHeader {
+    padding: 20px;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    border-bottom: 1px solid #e0e0e0;
+}
+
+.sidebarTitle {
+    font-size: 1.5rem;
+    font-weight: bold;
+    margin: 0;
+    color: #333333;
+}
+
+.toggleButton {
+    background: none;
+    border: none;
+    color: #666666;
+    cursor: pointer;
+    padding: 8px;
+    border-radius: 4px;
+    transition: background-color 0.2s;
+}
+
+.toggleButton:hover {
+    background-color: #f5f5f5;
+}
+
+.nav {
+    flex: 1;
+    padding: 20px 0;
+}
+
+.menuItem {
+    margin-bottom: 5px;
+}
+
+.menuButton {
+    width: 100%;
+    background: none;
+    border: none;
+    color: #333333;
+    padding: 12px 20px;
+    text-align: left;
+    cursor: pointer;
+    display: flex;
+    align-items: center;
+    gap: 12px;
+    transition: background-color 0.2s;
+    border-radius: 8px;
+    width: calc(100% - 20px);
+    margin: 0 10px;
+}
+
+.menuButton:hover {
+    background-color: hsla(231, 55%, 56%, 0.7);
+}
+
+.menuButtonActive {
+    background-color: hsl(231, 55%, 56%);
+    color: #ffffff;
+}
+
+.menuIcon {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    min-width: 20px;
+}
+
+.menuText {
+    font-size: 1rem;
+    font-weight: 500;
+}
+
+.subMenu {
+    background-color: #f8f9fa;
+    padding: 5px 0;
+    margin-right: 0;
+    border-radius: 0;
+}
+
+.subMenuButton {
+    width: 100%;
+    background: none;
+    border: none;
+    color: #666666;
+    padding: 10px 20px 10px 52px;
+    text-align: left;
+    cursor: pointer;
+    display: flex;
+    align-items: center;
+    gap: 8px;
+    transition: background-color 0.2s;
+    font-size: 0.9rem;
+    border-radius: 6px;
+    width: calc(100% - 30px);
+    margin: 2px 15px;
+}
+
+.subMenuButton:hover {
+    background-color: #e9ecef;
+    color: #333333;
+}
+
+.subMenuButtonActive {
+    background-color: #0056b3;
+    color: #ffffff;
+}
+
+.subMenuIcon {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    min-width: 18px;
+}
+
+.subMenuText {
+    font-size: 0.9rem;
+}
