+++ airbng-frontend/src/styles/admin/pages/AdminDashboard.module.css
@@ -1,0 +1,103 @@
+.contentCard {
+    background: white;
+    border-radius: 12px;
+    box-shadow: 0 2px 12px rgba(0, 0, 0, 0.1);
+    padding: 30px;
+    margin-bottom: 20px;
+}
+
+.contentTitle {
+    font-size: 1.8rem;
+    font-weight: 600;
+    color: #2c3e50;
+    margin-bottom: 20px;
+}
+
+.textGray {
+    color: #6c757d;
+}
+
+/* 통계 그리드 */
+.statsGrid {
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
+    gap: 20px;
+    margin-top: 25px;
+}
+
+.statCard {
+    background: #f8f9fa;
+    border: 1px solid #e9ecef;
+    border-radius: 8px;
+    padding: 20px;
+    text-align: center;
+}
+
+.statLabel {
+    font-size: 0.9rem;
+    color: #6c757d;
+    margin-bottom: 8px;
+    font-weight: 500;
+}
+
+.statValue {
+    font-size: 2rem;
+    font-weight: bold;
+    margin: 0;
+}
+
+/* 색상 */
+.textOrange {
+    color: #fd7e14;
+}
+
+.textGreen {
+    color: #28a745;
+}
+
+.textRed {
+    color: #dc3545;
+}
+
+/* 메뉴 그리드 */
+.menuGrid {
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
+    gap: 20px;
+    margin-top: 20px;
+}
+
+.menuCard {
+    background: #f8f9fa;
+    border: 1px solid #e9ecef;
+    border-radius: 8px;
+    padding: 24px;
+    text-align: center;
+    cursor: pointer;
+    transition: all 0.2s ease;
+}
+
+.menuCard:hover {
+    background: #e9ecef;
+    transform: translateY(-2px);
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+}
+
+.menuCardIcon {
+    font-size: 2rem;
+    color: #6c757d;
+    margin-bottom: 12px;
+}
+
+.menuCardTitle {
+    font-size: 1.1rem;
+    font-weight: 600;
+    margin-bottom: 8px;
+    color: #495057;
+}
+
+.menuCardDesc {
+    font-size: 0.9rem;
+    color: #6c757d;
+    line-height: 1.4;
+}
