--- airbng-frontend/src/styles/admin/pages/adminfirstpage.module.css
@@ -1,469 +1,0 @@
-.pageContainer {
-    width: 100%;
-    height: 100vh;
-    display: flex;
-    flex-direction: column;
-}
-
-/* 상단 헤더 스타일 */
-.topHeader {
-    background-color: #4561DB;
-    height: 60px;
-    display: flex;
-    align-items: center;
-    width: 100%;
-    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-    z-index: 1000;
-}
-
-.headerContent {
-    padding: 0 20px;
-    width: 100%;
-}
-
-.headerTitle {
-    color: #ffffff;
-    font-size: 1.5rem;
-    font-weight: bold;
-    margin: 0;
-}
-
-.container {
-    display: flex;
-    height: calc(100vh - 60px);
-    background-color: #f8f9fa;
-}
-
-/* 사이드바 - 하얀색으로 변경 */
-.sidebar {
-    background-color: #ffffff;
-    color: #333333;
-    transition: width 0.3s ease;
-    display: flex;
-    flex-direction: column;
-    border-right: 1px solid #e0e0e0;
-    box-shadow: 2px 0 4px rgba(0, 0, 0, 0.1);
-}
-
-.sidebarOpen {
-    width: 280px;
-}
-
-.sidebarClosed {
-    width: 80px;
-}
-
-.sidebarHeader {
-    padding: 20px;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    border-bottom: 1px solid #e0e0e0;
-}
-
-.sidebarTitle {
-    font-size: 1.5rem;
-    font-weight: bold;
-    margin: 0;
-    color: #333333;
-}
-
-.toggleButton {
-    background: none;
-    border: none;
-    color: #666666;
-    cursor: pointer;
-    padding: 8px;
-    border-radius: 4px;
-    transition: background-color 0.2s;
-}
-
-.toggleButton:hover {
-    background-color: #f5f5f5;
-}
-
-/* 네비게이션 */
-.nav {
-    flex: 1;
-    padding: 20px 0;
-}
-
-.menuItem {
-    margin-bottom: 5px;
-}
-
-.menuButton {
-    width: 100%;
-    background: none;
-    border: none;
-    color: #333333;
-    padding: 12px 20px;
-    text-align: left;
-    cursor: pointer;
-    display: flex;
-    align-items: center;
-    gap: 12px;
-    transition: background-color 0.2s;
-    border-radius: 8PX;
-    width: calc(100% - 20px);
-    margin: 0 10px;
-}
-
-.menuButton:hover {
-    background-color: hsla(231, 55%, 56%, 0.7);
-}
-
-.menuButtonActive {
-    background-color: hsl(231, 55%, 56%);
-    color: #ffffff;
-}
-
-.menuIcon {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    min-width: 20px;
-}
-
-.menuText {
-    font-size: 1rem;
-    font-weight: 500;
-}
-
-/* 서브메뉴 */
-.subMenu {
-    background-color: #f8f9fa;
-    padding: 5px 0;
-    margin-right: 0; /* 마진 제거 */
-    border-radius: 0; /* 둥근 모서리 제거 */
-}
-
-.subMenuButton {
-    width: 100%;
-    background: none;
-    border: none;
-    color: #666666;
-    padding: 10px 20px 10px 52px;
-    text-align: left;
-    cursor: pointer;
-    display: flex;
-    align-items: center;
-    gap: 8px;
-    transition: background-color 0.2s;
-    font-size: 0.9rem;
-    border-radius: 6px;
-    width: calc(100% - 30px);
-    margin: 2px 15px;
-}
-
-.subMenuButton:hover {
-    background-color: #e9ecef;
-    color: #333333;
-}
-
-.subMenuButtonActive {
-    background-color: #0056b3;
-    color: #ffffff;
-}
-
-.subMenuIcon {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    min-width: 18px;
-}
-
-.subMenuText {
-    font-size: 0.9rem;
-}
-
-/* 메인 컨텐츠 */
-.mainContent {
-    flex: 1;
-    overflow-y: auto;
-}
-
-.contentWrapper {
-    padding: 30px;
-    max-width: 1200px;
-    margin: 0 auto;
-}
-
-.contentCard {
-    background: white;
-    border-radius: 12px;
-    box-shadow: 0 2px 12px rgba(0, 0, 0, 0.1);
-    padding: 30px;
-    margin-bottom: 20px;
-}
-
-.contentTitle {
-    font-size: 1.8rem;
-    font-weight: 600;
-    color: #2c3e50;
-    margin-bottom: 20px;
-}
-
-.textGray {
-    color: #6c757d;
-}
-
-/* 통계 그리드 */
-.statsGrid {
-    display: grid;
-    grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
-    gap: 20px;
-    margin-top: 25px;
-}
-
-.statCard {
-    background: #f8f9fa;
-    border: 1px solid #e9ecef;
-    border-radius: 8px;
-    padding: 20px;
-    text-align: center;
-}
-
-.statLabel {
-    font-size: 0.9rem;
-    color: #6c757d;
-    margin-bottom: 8px;
-    font-weight: 500;
-}
-
-.statValue {
-    font-size: 2rem;
-    font-weight: bold;
-    margin: 0;
-}
-
-/* 색상 클래스 */
-.textOrange {
-    color: #fd7e14;
-}
-
-.textGreen {
-    color: #28a745;
-}
-
-.textRed {
-    color: #dc3545;
-}
-
-/* 메뉴 그리드 스타일 */
-.menuGrid {
-    display: grid;
-    grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
-    gap: 20px;
-    margin-top: 20px;
-}
-
-.menuCard {
-    background: #f8f9fa;
-    border: 1px solid #e9ecef;
-    border-radius: 8px;
-    padding: 24px;
-    text-align: center;
-    cursor: pointer;
-    transition: all 0.2s ease;
-}
-
-.menuCard:hover {
-    background: #e9ecef;
-    transform: translateY(-2px);
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-}
-
-.menuCardIcon {
-    font-size: 2rem;
-    color: #6c757d;
-    margin-bottom: 12px;
-}
-
-.menuCardTitle {
-    font-size: 1.1rem;
-    font-weight: 600;
-    margin-bottom: 8px;
-    color: #495057;
-}
-
-.menuCardDesc {
-    font-size: 0.9rem;
-    color: #6c757d;
-    line-height: 1.4;
-}
-
-/* 서브 타이틀 */
-.subTitle {
-    font-size: 1.3rem;
-    font-weight: 600;
-    color: #495057;
-    margin-bottom: 20px;
-}
-
-/* 매출 관련 스타일 */
-.salesGrid {
-    display: grid;
-    grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
-    gap: 20px;
-    margin-top: 20px;
-}
-
-.salesCard {
-    background: #f8f9fa;
-    border: 1px solid #e9ecef;
-    border-radius: 8px;
-    padding: 20px;
-    text-align: center;
-}
-
-.salesLabel {
-    font-size: 0.9rem;
-    color: #6c757d;
-    margin-bottom: 8px;
-    font-weight: 500;
-}
-
-.salesValue {
-    font-size: 1.5rem;
-    font-weight: bold;
-    color: #2c3e50;
-    margin: 0;
-}
-
-/* 보관소 리스트 */
-.storageList {
-    margin-top: 20px;
-}
-
-.storageItems {
-    display: flex;
-    flex-direction: column;
-    gap: 12px;
-}
-
-.storageItem {
-    background: #f8f9fa;
-    border: 1px solid #e9ecef;
-    border-radius: 8px;
-    padding: 16px 20px;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-}
-
-.storageName {
-    font-weight: 500;
-    color: #495057;
-}
-
-.storageInfo {
-    display: flex;
-    align-items: center;
-    gap: 12px;
-}
-
-.storageSales {
-    font-weight: 600;
-    color: #2c3e50;
-}
-
-.growthBadge {
-    padding: 4px 8px;
-    border-radius: 12px;
-    font-size: 0.8rem;
-    font-weight: 600;
-}
-
-.growthPositive {
-    background-color: #d4edda;
-    color: #155724;
-}
-
-.growthNegative {
-    background-color: #f8d7da;
-    color: #721c24;
-}
-
-/* 수익 관련 스타일 */
-.profitGrid {
-    display: grid;
-    grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
-    gap: 20px;
-    margin-top: 20px;
-}
-
-.profitCard {
-    background: #f8f9fa;
-    border: 1px solid #e9ecef;
-    border-radius: 8px;
-    padding: 20px;
-}
-
-.profitTitle {
-    font-size: 1.1rem;
-    font-weight: 600;
-    color: #495057;
-    margin-bottom: 16px;
-}
-
-.profitItems {
-    display: flex;
-    flex-direction: column;
-    gap: 12px;
-}
-
-.profitItem {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-}
-
-.profitLabel {
-    color: #6c757d;
-    font-size: 0.9rem;
-}
-
-.profitLabelBold {
-    color: #495057;
-    font-weight: 600;
-}
-
-.profitValue {
-    font-weight: 500;
-    color: #495057;
-}
-
-.profitValueBold {
-    font-weight: bold;
-    font-size: 1.1rem;
-}
-
-.divider {
-    border: none;
-    border-top: 1px solid #dee2e6;
-    margin: 8px 0;
-}
-
-.profitRateContainer {
-    text-align: center;
-}
-
-.profitRate {
-    font-size: 2.5rem;
-    font-weight: bold;
-    color: #28a745;
-    margin-bottom: 8px;
-}
-
-.profitRateLabel {
-    color: #6c757d;
-    font-size: 0.9rem;
-    margin-bottom: 4px;
-}
-
-.profitGrowth {
-    color: #28a745;
-    font-size: 0.8rem;
-    font-weight: 500;
-}
